made some adjustments
