+++ app/(auth)/signin/components/index.tsx
@@ -1,0 +1,3 @@
+export { SignInForm } from './signin-form'
+export { SignInHeader } from './signin-header'
+export { SignInFooter } from './signin-footer'
